Moved lettings model to new app
