+++ oc_lettings_site/settings.py
@@ -27,6 +27,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'profiles',
+    'lettings',
 ]
 
 MIDDLEWARE = [
